chore: more detailed api declarations
